Add order_ids to adjustments where they were missing
* Adds a migration to add order_ids to adjustments to pass new
validations.
* Change importer/order to use to_d instead of to_f
* Update specs

Conflicts:
	backend/spec/features/admin/orders/adjustments_spec.rb
	core/lib/spree/core/importer/order.rb

+++ core/lib/spree/core/importer/order.rb
@@ -123,7 +123,7 @@ module Spree
             begin
               adjustment = order.adjustments.build(
                 order:  order,
-                amount: a[:amount].to_f,
+                amount: a[:amount].to_d,
                 label:  a[:label]
               )
               adjustment.save!
